Update webpacked tests
