+++ .gitignore
@@ -6,3 +6,4 @@ _*
 node_modules
 playground
 *.svg
+test-bundle.js
