+++ test/index.html
@@ -1,3 +1,4 @@
+
 <!DOCTYPE html>
 <html lang="en">
     <head>
@@ -31,7 +32,22 @@
         <script src="../node_modules/expect.js/index.js"></script>
         <script src="../node_modules/mocha/mocha.js"></script>
         <script>mocha.setup("bdd");</script>
-        <script src="../.tmp/test-bundle.js"></script>
+        <script src="system.js"></script>
+        <script src="paper.js"></script>
+        <script src="ajax.js"></script>
+        <script src="snap-tests.js"></script>
+        <script src="class.js"></script>
+        <script src="primitives.js"></script>
+        <script src="colors.js"></script>
+        <script src="attrs.js"></script>
+        <script src="attradd.js"></script>
+        <script src="path.js"></script>
+        <script src="element.js"></script>
+        <script src="gradients.js"></script>
+        <script src="set.js"></script>
+        <script src="filter.js"></script>
+        <script src="matrix.js"></script>
+        <script src="mina.js"></script>
         <svg id="svgroot"></svg>
         <script>
             mocha.checkLeaks();
