+++ test/index.webpack.html
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8">
+        <meta name="copyright" content="Copyright (c) 2013 Adobe Systems Incorporated. All rights reserved.
+
+ Licensed under the Apache License, Version 2.0 (the “License”);
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an “AS IS” BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.">
+        <title>Snap Tests</title>
+        <style media="screen">
+            svg {
+                position: absolute;
+                top: 0;
+                left: 0;
+            }
+        </style>
+        <link rel="stylesheet" href="node_modules/mocha/mocha.css">
+    </head>
+    <body>
+        <div id="mocha"></div>
+        <script src="node_modules/expect.js/index.js"></script>
+        <script src="node_modules/mocha/mocha.js"></script>
+        <script>mocha.setup("bdd");</script>
+        <script src="test-bundle.js"></script>
+        <svg id="svgroot"></svg>
+        <script>
+            mocha.checkLeaks();
+            mocha.run();
+        </script>
+    </body>
+</html>
